+++ checkpoint/orbax/checkpoint/experimental/v1/_src/serialization/compatibility.py
@@ -14,8 +14,7 @@
 
 """Compatibility wrapper to help leaf handlers to work as V0 type_handlers."""
 
-import copy
-from typing import Generic, Sequence, Tuple, cast, get_args
+from typing import Any, Generic, Sequence, Tuple, cast, get_args
 from absl import logging
 from etils import epath
 import jax
@@ -32,6 +31,7 @@ from orbax.checkpoint.experimental.v1._src.path import types as path_types
 from orbax.checkpoint.experimental.v1._src.serialization import array_leaf_handler
 from orbax.checkpoint.experimental.v1._src.serialization import numpy_leaf_handler
 from orbax.checkpoint.experimental.v1._src.serialization import scalar_leaf_handler
+from orbax.checkpoint.experimental.v1._src.serialization import string_leaf_handler
 from orbax.checkpoint.experimental.v1._src.serialization import types
 from orbax.checkpoint.experimental.v1._src.synchronization import synchronization
 from orbax.checkpoint.experimental.v1._src.tree import types as tree_types
@@ -105,6 +105,7 @@ def _construct_deserialization_param(
     array_leaf_handler.AbstractArray
     | numpy_leaf_handler.AbstractNumpy
     | scalar_leaf_handler.AbstractScalar
+    | string_leaf_handler.AbstractString
     | None
 ]:
   """Constructs a DeserializationParam from a ParamInfo and RestoreArg."""
@@ -176,6 +177,9 @@ def _construct_deserialization_param(
   ):
     # scalar type
     value = restore_args.restore_type
+  elif issubclass(restore_args.restore_type, str):
+    # string type
+    value = str
   else:
     raise ValueError(
         'Unsupported restore_args: %s. Cannot construct DeserializationParam.'
@@ -271,14 +275,22 @@ def _convert_v1_metadata_to_v0(
     logging.vlog(1, 'NumpyMetadata: %s', ret)
     return ret
   elif isinstance(metadata, type):
-    # scalar type such as int, float or np.number
-    ret = value_metadata.ScalarMetadata(
-        name=name,
-        directory=directory,
-        dtype=metadata,
-    )
-    logging.vlog(1, 'ScalarMetadata: %s', ret)
-    return ret
+    if issubclass(metadata, str):
+      ret = value_metadata.StringMetadata(
+          name=name,
+          directory=directory,
+      )
+      logging.vlog(1, 'StringMetadata: %s', ret)
+      return ret
+    else:
+      # scalar type such as int, float or np.number
+      ret = value_metadata.ScalarMetadata(
+          name=name,
+          directory=directory,
+          dtype=metadata,
+      )
+      logging.vlog(1, 'ScalarMetadata: %s', ret)
+      return ret
   elif isinstance(metadata, (np.dtype, jnp.dtype)):
     # scalar type as np.dtype instances
     ret = value_metadata.ScalarMetadata(
@@ -436,55 +448,36 @@ class CompatibleTypeHandler(
       return None
 
 
-def get_compatible_type_handler_registry(
-    context: context_lib.Context | None = None,
-    type_handler_registry: type_handlers_v0.TypeHandlerRegistry | None = None,
-) -> type_handlers_v0.TypeHandlerRegistry:
-  """Returns a V0 type handler registry that using v1 leaf handlers.
-
-  This is a helper function to setup a v0 type handler registry that will be
-  registered all existing v1 leaf handlers.
+def get_v0_type_handler_registry(
+    leaf_handler_registry: types.LeafHandlerRegistry,
+):
+  """Returns a v0 type handler registry based on the `leaf_handler_registry`.
 
   Args:
-    context: The context to that will be passed to create leaf handlers.
-    type_handler_registry: The v0 type handler registry to use. If not provided,
-      a new registry will be created.
-
-  Returns:
-    The v0 type handler registry that using standard v1 leaf handlers.
+    leaf_handler_registry: The LeafHandlerRegistry to be used to create a v0
+      type handler registry.
   """
-  if type_handler_registry is None:
-    type_handler_registry = copy.deepcopy(
-        type_handlers_v0.GLOBAL_TYPE_HANDLER_REGISTRY
-    )
 
-  type_handler_registry.add(
-      jax.Array,
-      CompatibleTypeHandler(
-          array_leaf_handler.ArrayLeafHandler(context=context),
-          typestr=type_handlers_v0.JAX_ARRAY_TYPE_STR,
-      ),
-      override=True,
-      ignore_warnings=True,
-  )
-  type_handler_registry.add(
-      np.ndarray,
-      CompatibleTypeHandler(
-          numpy_leaf_handler.NumpyLeafHandler(context=context),
-          typestr='np.ndarray',
-      ),
-      override=True,
-      ignore_warnings=True,
-  )
+  def _get_typestr(leaf_type: Any) -> str:
+    if leaf_type == jax.Array:
+      return type_handlers_v0.JAX_ARRAY_TYPE_STR
+    elif leaf_type == np.ndarray:
+      return 'np.ndarray'
+    elif leaf_type in (int, float, bytes, np.number):
+      return 'scalar'
+    elif leaf_type == str:
+      return 'string'
+    else:
+      return f'{leaf_type!r}'
 
-  for scalar_type in (int, float, bytes, np.number):
-    type_handler_registry.add(
-        scalar_type,
+  # register standardard v1 leaf handlers to the v0 type handler registry.
+  handlers = []
+  for leaf_type, _, leaf_handler in leaf_handler_registry.get_all():
+    handlers.append((
+        leaf_type,
         CompatibleTypeHandler(
-            scalar_leaf_handler.ScalarLeafHandler(context=context),
-            typestr='scalar',
+            leaf_handler,
+            typestr=_get_typestr(leaf_type),
         ),
-        override=True,
-        ignore_warnings=True,
-    )
-  return type_handler_registry
+    ))
+  return type_handlers_v0.create_type_handler_registry(*handlers)