+++ checkpoint/CHANGELOG.md
@@ -7,12 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- #v1 Add LeafHandlerRegistry and make it as default for V1 codes
+
 ## [0.11.19] - 2025-07-08
 
 ### Added
 
 - #v1 Add dummy `orbax.checkpoint` file to help identify checkpoints.
-- #v1 Add LeafHandlerRegistry
 - #v1 Add StringLeafHandler
 
 ### Changed